Updated Rucksack

+++ com.konsteyyy.git.uni.dea/src/com/konsteyyy/git/uni/dea/rucksack/Item.java
@@ -1,23 +1,23 @@
 package com.konsteyyy.git.uni.dea.rucksack;
 
 public class Item implements Comparable<Item>{
-	private double weight;
-	private double value;
+	private int weight;
+	private int value;
 	private double ratio;
 	
-	public Item(double value, double weight) {
+	public Item(int value, int weight) {
 		this.value = value;
 		this.weight = weight;
-		this.ratio = value/weight;
+		this.ratio = (double) value/weight;
 	}
 
-	public double getWeight() {return weight;}
+	public int getWeight() {return weight;}
 
-	public void setWeight(double weight) {this.weight = weight;}
+	public void setWeight(int weight) {this.weight = weight;}
 
-	public double getValue() {return value;}
+	public int getValue() {return value;}
 
-	public void setValue(double value) {this.value = value;}
+	public void setValue(int value) {this.value = value;}
 
 	public double getRatio() {return ratio;}
